Replace futures with normal callbacks
Futures weren't implemented in a clean way. For example,
in the dialog middleware, we would just ignore the return
value of the next() call and instead return our own future.
So it depended on the order of the middlerwares, if the
correct future would be passed back to the caller.